% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_capa.R
\name{gerar_capa}
\alias{gerar_capa}
\title{Gerar Capa do Relatório}
\usage{
gerar_capa(titulo, gerente, assessores, bloco = FALSE)
}
\arguments{
\item{titulo}{String com o título do relatório}

\item{gerente}{String com o nome do gerente da consultoria}

\item{assessores}{Vetor de caracteres com os assessores da consultoria}
}
\value{
character
}
\description{
Retorna string com código em LaTeX para gerar a capa do relatório de consultorias
}
\examples{
 gerar_capa("Análise de alguma coisa", "Samuel Vianna Quintanilha", c("Ellen Beatriz Balan", "Larissa Rei Dos Santos"))

}
