% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodologia.R
\name{met_opts}
\alias{met_opts}
\title{Metodologias Disponíveis}
\usage{
met_opts()
}
\value{
character
}
\description{
Lista todas as metodologias disponíveis para ser utilizada em estatsr::metodologia
}
\examples{
 met_opts()

}
