% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projeto.R
\name{projeto}
\alias{projeto}
\title{Projeto}
\usage{
projeto(nome)
}
\arguments{
\item{nome}{Nome da pasta do projeto a ser criada}
}
\description{
Cria diretórios de projetos para consultorias e copia arquivos base de relatórios
}
\examples{
 projeto("056-Eduardo")

}
