% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gerar_capa.R
\name{gerar_capa}
\alias{gerar_capa}
\title{Gerar Capa do Relatório}
\usage{
gerar_capa(
  subtitulo,
  gerente,
  assessores,
  titulo = "Relatório Final",
  bloco = FALSE
)
}
\arguments{
\item{gerente}{String com o nome do gerente da consultoria}

\item{assessores}{Vetor de caracteres com os assessores da consultoria}

\item{titulo}{String com o título do relatório}

\item{bloco}{Valor lógico que se definido como TRUE faz com que a função gere um texto para bloco de código, e se definido como FALSE gera texto para R inline}
}
\value{
character
}
\description{
Retorna string com código em LaTeX para gerar a capa do relatório de consultorias
}
\examples{
 gerar_capa("Análise de alguma coisa", "Samuel Vianna Quintanilha", c("Ellen Beatriz Balan", "Larissa Rei Dos Santos"))

}
