% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodologia.R
\name{metodologia}
\alias{metodologia}
\title{Metodologias}
\usage{
metodologia(metodologias)
}
\arguments{
\item{metodologias}{Vetor de caracteres com os nomes das metodologias a serem retornadas (Strings não devem conter acentos)}
}
\value{
character
}
\description{
Retorna a ou as metodologias pedidas
}
\examples{
 metodologia("teste de hipotese")
 metodologia(c("Teste de hipotese", "Teste t"))

}
