% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodologia.R
\name{metodologia}
\alias{metodologia}
\title{Metodologias}
\usage{
metodologia(metodologias)
}
\arguments{
\item{metodologias}{Vetor de caracteres com os nomes das metodologias a serem retornadas (Strings não devem conter acentos)}
}
\description{
Retorna a ou as metodologias pedidas
}
